proc: Always report hard reset to userspace in /proc/cmdline
The stock Asus ROM attempts to issue a hard reset whenever it detects a
boot reason other than hard reset ("hard"). The bootloader passes
"warm-reset" when we issue a warm reset to reboot the device, which
triggers this condition.

We need warm resets to preserve logs after a panic, and there's no
known downsides other than this artificial limitation, so patch the
command line shown to userspace to make it happy.

Co-authored-by: Danny Lin <danny@kdrag0n.dev>
Co-authored-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: kingbri <bdashore3@gmail.com>